reorganize componenents dir
