+++ package.json
@@ -31,7 +31,7 @@
     "babel-cli": "^6.24.0",
     "babel-core": "^6.24.0",
     "babel-loader": "^7.1.1",
-    "babel-plugin-universal-import": "^1.2.2",
+    "babel-plugin-universal-import": "^1.3.4",
     "babel-polyfill": "^6.23.0",
     "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.23.0",
