+++ src/components/App/App.css
@@ -3,15 +3,22 @@
 }
 
 .App--slideover-open .StickyNav,
-.App--slideover-open .Main {
+.App--slideover-open .App__Content {
   transform: translateX(255px);
 }
 
 .StickyNav,
-.Main {
+.App__Content {
   transition: transform 0.4s;
 }
 
 .App--slideover-transitioning {
   overflow: hidden;
+}
+
+.App__Content {
+  padding-top: 52px;
+  position: relative;
+  background: #fff;
+  min-height: 100vh;
 }
