--- src/components/Loading.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export default () => <div><div /></div>;
