--- src/components/Main.css
@@ -1,6 +1,0 @@
-.Main {
-  padding-top: 52px;
-  position: relative;
-  background: #fff;
-  min-height: 100vh;
-}
