--- src/components/NotFound.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export default () => <div>PAGE NOT FOUND - 404</div>;
