+++ src/components/Page/components/Audio/Container.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import type { State } from 'type';
 import Document from 'classes/Document';
 import Edition from 'classes/Edition';
-import AudioPage from './AudioPage';
+import AudioPage from './';
 
 const mapStateToProps = (state: State): Object => {
   const { location: { payload: { friendSlug, documentSlug, editionType } } } = state;
