+++ src/components/Page/components/Friend/Container.js
@@ -1,7 +1,7 @@
 // @flow
 import { connect } from 'react-redux';
 import type { State } from 'type';
-import FriendPage from './FriendPage';
+import FriendPage from './';
 
 const mapStateToProps = (state: State): Object => {
   const { location: { payload: { friendSlug } } } = state;
