+++ src/components/Page/components/Friend/components/DocumentTeaser/index.js
