+++ src/components/Page/components/Home/Container.js
@@ -1,0 +1,4 @@
+// @flow
+import Home from './';
+
+export default Home;
