+++ src/components/Page/components/NotFound/Container.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export default () => <h1>404</h1>;
