+++ src/components/Page/index.js
@@ -2,12 +2,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import universal from 'react-universal-component';
-import Loading from './Loading';
 
 // $FlowFixMe
-const UniversalComponent = universal(({ page }) => import(`./${page}`), {
-  loading: Loading,
-});
+const UniversalComponent = universal(({ page }) => import(`./components/${page}/Container`));
 
 const Page = ({ page }: { page: string }) => (
   <UniversalComponent page={page} />
